+++ src/main/kotlin/org/coner/snoozle/db/LiteralRecordDefinition.kt
@@ -1,10 +1,14 @@
 package org.coner.snoozle.db
 
 import java.util.*
+import kotlin.reflect.KClass
 
-abstract class LiteralRecordDefinition<R : Record<K>, K : Key> : RecordDefinition<R, K>() {
+abstract class LiteralRecordDefinition<K : Key, R : Record<K>>(
+        keyClass: KClass<K>,
+        recordClass: KClass<R>
+) : RecordDefinition<K, R>(keyClass, recordClass) {
 
-    operator fun String.div(uuidExtractor: K.() -> UUID): MutableList<PathPart<R, K, *>> {
+    operator fun String.div(uuidExtractor: K.() -> UUID): MutableList<PathPart<K, R, *>> {
         return mutableListOf(
                 PathPart.StringValue(this),
                 PathPart.DirectorySeparator(),
@@ -12,30 +16,30 @@ abstract class LiteralRecordDefinition<R : Record<K>, K : Key> : RecordDefinitio
         )
     }
 
-    operator fun MutableList<PathPart<R, K, *>>.div(part: String): MutableList<PathPart<R, K, *>> {
+    operator fun MutableList<PathPart<K, R, *>>.div(part: String): MutableList<PathPart<K, R, *>> {
         add(PathPart.DirectorySeparator())
         add(PathPart.StringValue(part))
         return this
     }
 
-    operator fun MutableList<PathPart<R, K, *>>.div(uuidExtractor: K.() -> UUID): MutableList<PathPart<R, K, *>> {
+    operator fun MutableList<PathPart<K, R, *>>.div(uuidExtractor: K.() -> UUID): MutableList<PathPart<K, R, *>> {
         add(PathPart.DirectorySeparator())
         add(PathPart.UuidVariable(uuidExtractor))
         return this
     }
 
-    operator fun MutableList<PathPart<R, K, *>>.plus(extension: String): MutableList<PathPart<R, K, *>> {
+    operator fun MutableList<PathPart<K, R, *>>.plus(extension: String): MutableList<PathPart<K, R, *>> {
         add(PathPart.StringValue(extension))
         return this
     }
 
-    operator fun MutableList<PathPart<R, K, *>>.div(stringExtractor: StringArgumentExtractor<K>): MutableList<PathPart<R, K, *>> {
+    operator fun MutableList<PathPart<K, R, *>>.div(stringExtractor: StringArgumentExtractor<K>): MutableList<PathPart<K, R, *>> {
         add(PathPart.DirectorySeparator())
         add(PathPart.StringVariable(stringExtractor::extract))
         return this
     }
 
-    operator fun MutableList<PathPart<R, K, *>>.plus(stringExtractor: StringArgumentExtractor<K>): MutableList<PathPart<R, K, *>> {
+    operator fun MutableList<PathPart<K, R, *>>.plus(stringExtractor: StringArgumentExtractor<K>): MutableList<PathPart<K, R, *>> {
         add(PathPart.StringVariable(stringExtractor::extract))
         return this
     }