+++ src/main/kotlin/org/coner/snoozle/db/Record.kt
@@ -1,5 +1,3 @@
 package org.coner.snoozle.db
 
-interface Record<K : Key> {
-    val key: K
-}
+interface Record<K : Key>