+++ src/main/kotlin/org/coner/snoozle/db/RecordDefinition.kt
@@ -1,6 +1,14 @@
 package org.coner.snoozle.db
 
-abstract class RecordDefinition<R : Record<K>, K : Key> {
+import kotlin.reflect.KClass
+
+abstract class RecordDefinition<K : Key, R : Record<K>>(
+        val keyClass: KClass<K>,
+        val recordClass: KClass<R>
+) {
+
+    var path: List<PathPart<K, R, *>> = mutableListOf()
+    var keyFromPath: (KeyMapper<K, R>.RelativeRecordContext.() -> K)? = null
 
     abstract class PathArgumentExtractor<K : Key, P>(private val fn: K.() -> P) {
         fun extract(key: K): P {