+++ src/main/kotlin/org/coner/snoozle/db/TypesRegistry.kt
@@ -1,11 +1,12 @@
 package org.coner.snoozle.db
 
-import com.fasterxml.jackson.core.type.TypeReference
 import com.fasterxml.jackson.databind.ObjectMapper
 import org.coner.snoozle.db.blob.Blob
 import org.coner.snoozle.db.blob.BlobDefinition
 import org.coner.snoozle.db.blob.BlobResource
-import org.coner.snoozle.db.entity.*
+import org.coner.snoozle.db.entity.Entity
+import org.coner.snoozle.db.entity.EntityDefinition
+import org.coner.snoozle.db.entity.EntityResource
 import java.nio.file.Path
 import kotlin.reflect.KClass
 
@@ -15,50 +16,53 @@ class TypesRegistry(
         op: TypesRegistry.() -> Unit
 ) {
     val entityResources = mutableMapOf<KClass<*>, EntityResource<*, *>>()
-    val versionedEntityResources = mutableMapOf<KClass<*>, VersionedEntityResource<*, *>>()
     val blobResources = mutableMapOf<KClass<*>, BlobResource<*>>()
 
     init {
         this.op()
     }
 
-    inline fun <reified E : Entity<K>, reified K : Key> entity(op: EntityDefinition<E, K>.() -> Unit) {
-        val entityDefinition = EntityDefinition<E, K>().apply(op)
+    inline fun <reified K : Key, reified E : Entity<K>> entity(op: EntityDefinition<K, E>.() -> Unit) {
+        val entityDefinition = EntityDefinition<K, E>(
+                keyClass = K::class,
+                entityClass = E::class
+        ).apply(op)
+        val pathfinder = Pathfinder(
+                root = root,
+                pathParts = entityDefinition.path
+        )
         entityResources[E::class] = EntityResource(
                 root = root,
                 definition = entityDefinition,
                 reader = objectMapper.readerFor(E::class.java),
                 writer = objectMapper.writerFor(E::class.java),
-                path = Pathfinder(
-                        root = root,
-                        pathParts = entityDefinition.path
-                ),
-        )
-    }
-
-    inline fun <reified VE : VersionedEntity<K>, K : Key> versionedEntity(op: VersionedEntityDefinition<VE, K>.() -> Unit) {
-        val versionedEntityDefinition = VersionedEntityDefinition<VE, K>().apply(op)
-        val jacksonTypeReference = object : TypeReference<VersionedEntityContainer<VE, K>>() { }
-        versionedEntityResources[VE::class] = VersionedEntityResource(
-                root = root,
-                definition = versionedEntityDefinition,
-                reader = objectMapper.readerFor(jacksonTypeReference),
-                writer = objectMapper.writerFor(jacksonTypeReference),
-                path = VersionedEntityPathfinder(
-                        root = root,
-                        pathParts = versionedEntityDefinition.path
+                pathfinder = pathfinder,
+                keyMapper = KeyMapper(
+                        definition = entityDefinition,
+                        pathfinder = pathfinder,
+                        relativeRecordFn = requireNotNull(entityDefinition.keyFromPath),
+                        instanceFn = requireNotNull(entityDefinition.keyFromEntity)
                 )
         )
     }
 
     inline fun <reified B : Blob> blob(op: BlobDefinition<B>.() -> Unit) {
-        val blobDefinition = BlobDefinition<B>().apply(op)
+        val definition = BlobDefinition(
+                blobClass = B::class
+        ).apply(op)
+        val pathfinder = Pathfinder(
+                root = root,
+                pathParts = definition.path
+        )
         blobResources[B::class] = BlobResource(
                 root = root,
-                definition =  blobDefinition,
-                path = Pathfinder(
-                        root = root,
-                        pathParts = blobDefinition.path
+                definition =  definition,
+                pathfinder = pathfinder,
+                keyMapper = KeyMapper(
+                        definition = definition,
+                        pathfinder = pathfinder,
+                        relativeRecordFn = requireNotNull(definition.keyFromPath),
+                        instanceFn = { this as B }
                 )
         )
     }