+++ src/main/kotlin/org/coner/snoozle/db/blob/Blob.kt
@@ -3,8 +3,4 @@ package org.coner.snoozle.db.blob
 import org.coner.snoozle.db.Key
 import org.coner.snoozle.db.Record
 
-interface Blob : Record<Blob>, Key {
-
-    override val key: Blob
-        get() = this
-}
+interface Blob : Record<Blob>, Key