+++ src/main/kotlin/org/coner/snoozle/db/entity/EntityEvent.kt
@@ -3,7 +3,7 @@ package org.coner.snoozle.db.entity
 import org.coner.snoozle.db.Key
 import java.util.*
 
-data class EntityEvent<E : Entity<K>, K : Key>(val state: State, val id: UUID, val entity: E? = null) {
+data class EntityEvent<K : Key, E : Entity<K>>(val state: State, val key: K, val entity: E? = null) {
     enum class State {
         EXISTS,
         DELETED,