+++ src/main/kotlin/org/coner/snoozle/db/entity/EntityResource.kt
@@ -4,41 +4,41 @@ import com.fasterxml.jackson.databind.ObjectReader
 import com.fasterxml.jackson.databind.ObjectWriter
 import io.reactivex.Observable
 import org.coner.snoozle.db.Key
+import org.coner.snoozle.db.KeyMapper
 import org.coner.snoozle.db.Pathfinder
-import org.coner.snoozle.util.nameWithoutExtension
-import org.coner.snoozle.util.uuid
 import org.coner.snoozle.util.watch
 import java.nio.file.*
 import java.util.function.Predicate
 import java.util.stream.Stream
 
-class EntityResource<E : Entity<K>, K : Key> constructor(
+class EntityResource<K : Key, E : Entity<K>> constructor(
         private val root: Path,
-        internal val definition: EntityDefinition<E, K>,
+        internal val definition: EntityDefinition<K, E>,
         private val reader: ObjectReader,
         private val writer: ObjectWriter,
-        private val path: Pathfinder<E, K>,
-        private val key: EntityKeyParser<E, K>
+        private val pathfinder: Pathfinder<K, E>,
+        private val keyMapper: KeyMapper<K, E>
 ) {
 
     fun get(key: K): E {
-        val entityPath = path.findRecord(key)
+        val entityPath = pathfinder.findRecord(key)
         val file = root.resolve(entityPath)
         return read(file)
     }
 
     fun put(entity: E) {
-        val entityParentPath = path.findListingByRecord(entity)
-        val parent = root.resolve(entityParentPath)
-        if (!Files.exists(parent)) {
+        val key = keyMapper.fromInstance(entity)
+        val relativeRecord = pathfinder.findRecord(key)
+        val absoluteParent = root.resolve(relativeRecord.parent)
+        if (!Files.exists(absoluteParent)) {
             try {
-                Files.createDirectories(parent)
+                Files.createDirectories(absoluteParent)
             } catch (t: Throwable) {
-                val message = "Failed to create parent folder for ${definition::class.java.simpleName} $entityParentPath"
+                val message = "Failed to create parent folder for ${definition.recordClass.simpleName} ${relativeRecord.parent}"
                 throw EntityIoException.WriteFailure(message, t)
             }
         }
-        val file = root.resolve(path.findRecord(entity))
+        val file = root.resolve(relativeRecord)
         write(file, entity)
     }
 
@@ -64,47 +64,54 @@ class EntityResource<E : Entity<K>, K : Key> constructor(
         Files.move(tempFile, destination, StandardCopyOption.REPLACE_EXISTING, StandardCopyOption.ATOMIC_MOVE)
     }
 
-    fun delete(entity: E) {
-        val file = root.resolve(path.findRecord(entity))
+    fun delete(key: K) {
+        val file = root.resolve(pathfinder.findRecord(key))
         Files.delete(file)
     }
 
-    fun streamAll(keyFilter: Predicate<K>? = null): Stream<E> {
-        val allPathsMappedToKeys = path.streamAll()
-                .map { recordPath: Path -> key.parse(recordPath) }
+    fun stream(keyFilter: Predicate<K>? = null): Stream<E> {
+        val allPathsMappedToKeys = pathfinder.streamAll()
+                .map { recordPath: Path -> keyMapper.fromRelativeRecord(recordPath) }
         val allKeysForRead = keyFilter?.let { allPathsMappedToKeys.filter(it) } ?: allPathsMappedToKeys
-        return allKeysForRead.map { read(path.findRecord(it)) }
+        return allKeysForRead.map { read(pathfinder.findRecord(it)) }
     }
 
-    fun watchListing(): Observable<EntityEvent<E, K>> {
+    fun watch(keyFilter: Predicate<K>? = null): Observable<EntityEvent<K, E>> {
         return root.watch(recursive = true)
-                .filter { path.isRecord(root.relativize(it.context() as Path)) }
+                .map { event: WatchEvent<*> ->
+                    val file = event.context() as Path
+                    PreReadWatchEventPayload(
+                            event = event,
+                            file = file,
+                            key = keyMapper.fromRelativeRecord(root.relativize(file))
+                    )
+                }
+                .filter { pathfinder.isRecord(it.file) && keyFilter?.test(it.key) != false }
                 .map {
-                    val file = it.context() as Path
-                        val entity = if (it.kind() != StandardWatchEventKinds.ENTRY_DELETE
-                                && Files.exists(file)
-                                && Files.size(file) > 0
-                        ) {
-                            try {
-                                read(file)
-                            } catch (t: Throwable) {
-                                null
-                            }
-                        } else {
+                    val entity = if (it.event.kind() != StandardWatchEventKinds.ENTRY_DELETE
+                            && Files.exists(it.file)
+                            && Files.size(it.file) > 0
+                    ) {
+                        try {
+                            read(it.file)
+                        } catch (t: Throwable) {
                             null
                         }
-                        it to entity
+                    } else {
+                        null
+                    }
+                    PostReadWatchEventPayload(it.event, it.key, entity)
                 }
                 .filter {
-                    when (it.first.kind()) {
-                        StandardWatchEventKinds.ENTRY_MODIFY, StandardWatchEventKinds.ENTRY_CREATE -> it.second != null
-                        StandardWatchEventKinds.ENTRY_DELETE -> it.second == null
+                    when (it.event.kind()) {
+                        StandardWatchEventKinds.ENTRY_MODIFY, StandardWatchEventKinds.ENTRY_CREATE -> it.entity != null
+                        StandardWatchEventKinds.ENTRY_DELETE -> it.entity == null
                         StandardWatchEventKinds.OVERFLOW -> true
                         else -> false
                     }
                 }
                 .map {
-                    val state = when(it.first.kind()) {
+                    val state = when(it.event.kind()) {
                         StandardWatchEventKinds.ENTRY_MODIFY, StandardWatchEventKinds.ENTRY_CREATE -> EntityEvent.State.EXISTS
                         StandardWatchEventKinds.ENTRY_DELETE -> EntityEvent.State.DELETED
                         StandardWatchEventKinds.OVERFLOW -> EntityEvent.State.OVERFLOW
@@ -112,9 +119,21 @@ class EntityResource<E : Entity<K>, K : Key> constructor(
                     }
                     EntityEvent(
                             state = state,
-                            id = uuid((it.first.context() as Path).nameWithoutExtension),
-                            entity = it.second
+                            key = it.key,
+                            entity = it.entity
                     )
                 }
     }
+
+    private class PreReadWatchEventPayload<K : Key>(
+        val event: WatchEvent<*>,
+        val file: Path,
+        val key: K
+    )
+
+    private class PostReadWatchEventPayload<K : Key, E : Entity<K>>(
+            val event: WatchEvent<*>,
+            val key: K,
+            val entity: E?
+    )
 }