+++ src/main/kotlin/org/coner/snoozle/db/entity/VersionedEntity.kt
@@ -1,6 +1,0 @@
-package org.coner.snoozle.db.entity
-
-import org.coner.snoozle.db.Key
-
-interface VersionedEntity<K : Key> : Entity<K> {
-}