+++ src/test/kotlin/org/coner/snoozle/db/sample/Gadget.kt
@@ -1,7 +1,7 @@
 package org.coner.snoozle.db.sample
 
 import org.coner.snoozle.db.Key
-import org.coner.snoozle.db.entity.VersionedEntity
+import org.coner.snoozle.db.entity.Entity
 import java.time.ZonedDateTime
 import java.util.*
 
@@ -9,9 +9,10 @@ data class Gadget(
         val id: UUID = UUID.randomUUID(),
         var name: String? = null,
         var silly: ZonedDateTime? = null
-) : VersionedEntity<GadgetKey> {
+) : Entity<Gadget.Key> {
 
-    override val key by lazy { GadgetKey(id = id) }
+    override val key by lazy { Key(id = id) }
+
+    data class Key(val id: UUID) : org.coner.snoozle.db.Key
+
 }
-
-data class GadgetKey(val id: UUID) : Key